test-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --constraint=requirements.txt test-requirements.in
#
asttokens==2.4.1
    # via stack-data
bcrypt==4.2.0
    # via paramiko
cachetools==5.5.0
    # via google-auth
certifi==2024.7.4
    # via
    #   kubernetes
    #   requests
cffi==1.16.0
    # via
    #   -c requirements.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
codespell==2.3.0
    # via -r test-requirements.in
coverage[toml]==7.6.1
    # via -r test-requirements.in
cryptography==43.0.0
    # via
    #   -c requirements.txt
    #   paramiko
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
executing==2.0.1
    # via stack-data
google-auth==2.34.0
    # via kubernetes
hvac==2.3.0
    # via juju
idna==3.7
    # via requests
iniconfig==2.0.0
    # via
    #   -c requirements.txt
    #   pytest
ipdb==0.13.13
    # via pytest-operator
ipython==8.26.0
    # via ipdb
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via pytest-operator
juju==3.5.2.0
    # via
    #   -r test-requirements.in
    #   pytest-operator
kubernetes==30.1.0
    # via juju
macaroonbakery==1.3.4
    # via juju
markupsafe==2.1.5
    # via jinja2
matplotlib-inline==0.1.7
    # via ipython
mypy-extensions==1.0.0
    # via typing-inspect
nodeenv==1.9.1
    # via pyright
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
ops==2.15.0
    # via
    #   -c requirements.txt
    #   ops-scenario
ops-scenario==6.0.1
    # via
    #   -c requirements.txt
    #   -r test-requirements.in
packaging==23.2
    # via
    #   -c requirements.txt
    #   juju
    #   pytest
paramiko==3.4.1
    # via juju
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pluggy==1.5.0
    # via
    #   -c requirements.txt
    #   pytest
prompt-toolkit==3.0.47
    # via ipython
protobuf==5.27.3
    # via macaroonbakery
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.0
    # via
    #   juju
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.21
    # via
    #   -c requirements.txt
    #   cffi
pygments==2.18.0
    # via ipython
pymacaroons==0.13.0
    # via macaroonbakery
pynacl==1.5.0
    # via
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
pyrfc3339==1.1
    # via
    #   juju
    #   macaroonbakery
pyright==1.1.377
    # via -r test-requirements.in
pytest==8.3.2
    # via
    #   -c requirements.txt
    #   -r test-requirements.in
    #   pytest-asyncio
    #   pytest-operator
pytest-asyncio==0.21.2
    # via
    #   -r test-requirements.in
    #   pytest-operator
pytest-operator==0.36.0
    # via -r test-requirements.in
python-dateutil==2.9.0.post0
    # via kubernetes
pytz==2024.1
    # via pyrfc3339
pyyaml==6.0.1
    # via
    #   -c requirements.txt
    #   juju
    #   kubernetes
    #   ops
    #   ops-scenario
    #   pytest-operator
requests==2.32.3
    # via
    #   hvac
    #   kubernetes
    #   macaroonbakery
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rsa==4.9
    # via google-auth
ruff==0.6.1
    # via -r test-requirements.in
six==1.16.0
    # via
    #   asttokens
    #   kubernetes
    #   macaroonbakery
    #   pymacaroons
    #   python-dateutil
stack-data==0.6.3
    # via ipython
toposort==1.10
    # via juju
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.9.0
    # via
    #   -c requirements.txt
    #   typing-inspect
typing-inspect==0.9.0
    # via juju
urllib3==2.2.2
    # via
    #   kubernetes
    #   requests
wcwidth==0.2.13
    # via prompt-toolkit
websocket-client==1.7.0
    # via
    #   -c requirements.txt
    #   kubernetes
    #   ops
websockets==13.0
    # via juju
